turned off mdash generation
In the long run, we need to figure out what's making Apache not output
&mdash; correctly. We've now tried both outputting the entity and the
utf-8 character, but these are both emitted as something else by Apache.
As far as we can see, all encoding directives are set to utf-8, both in
the HTTP header and in the <head> element in the HTML. Very strange.